requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
Pillow
PyQt5
PyJWT
